Add a grid::parse(std::istream&) overload
This will read chars from the given stream until either it has taken 81
of them or the stream is exhausted.

Annoyingly, we can’t do this in a very Range-y way, but need to resort
to an ordinary for-loop since ranges::view::remove_if will eat up the
next token which breaks trying to read multiple sudoku from a single
file.