Add operator< and friends for grids
Not 100% sure that this is a good idea. Yes, we want to be able to
place grids in std::sets, but I’m not sure that there is a natural
ordering for grids (or at least, grids with equal numbers of empty
cells); perhaps specialising std::less would be more appropriate?