Use a simpler unsolvable grid in tests
…because otherwise the tests take ~30 seconds to run in release mode on
GCC, and several minutes in debug mode. With the simpler grid, they run
in less than a second.

However, the reason this is simpler is because we can a contradiction
in the very first call to assign() in the solver — that is, we never
actually get as far as trying to *solve* anything. So I’m not 100% sure
about it. I feel like I ought to have a “— include-expensive” flag in
the test options which enables the other unsolvable test, for use with
CI.